refactor!: move operations into own module
