+++ workspaces/src/types/account.rs
@@ -5,7 +5,7 @@ use near_primitives::views::AccountView;
 use crate::types::{AccountId, Balance, InMemorySigner};
 use crate::{CryptoHash, Network, Worker};
 
-use super::transaction::{CallTransaction, CreateAccountTransaction, Transaction};
+use crate::operations::{CallTransaction, CreateAccountTransaction, Transaction};
 use crate::result::{CallExecution, CallExecutionDetails, ViewResultDetails};
 
 pub struct Account {
